feat: extract topnav
